auto logout when user don't interact for 10 min

+++ src/provider/helper/axios.js
@@ -132,7 +132,6 @@ import { REFRESH_TOKEN } from "../../../utils/urls.js";
 import { Route } from "react-router-dom";
 
 export const checkForTokenExpiredError = (error) => {
-  console.log("🚀 ~ checkForTokenExpiredError ~ error:", error);
   if (error.message === "Network Error") {
     document.write("Server Down");
   } else {
@@ -151,18 +150,22 @@ export const createWithAuth = () => {
 };
 
 const refreshTokenAndSetAuth = async (callback) => {
-  console.log("🚀 ~ refreshTokenAndSetAuth ~ callback:", callback);
   const refreshToken = localStorage.getItem("refreshToken");
-  const response = await refreshUserToken(REFRESH_TOKEN, {
-    refresh: refreshToken,
-  });
-  const { access } = response?.data;
-  if (response.status == 200) {
-    localStorage.setItem("token", access);
-    return callback();
+  if (refreshToken) {
+    const response = await refreshUserToken(REFRESH_TOKEN, {
+      refresh: refreshToken,
+    });
+    const { access } = response?.data;
+    if (response.status == 200) {
+      localStorage.setItem("token", access);
+      return callback();
+    } else {
+      localStorage.clear();
+      window.location.href = "/login";
+      Route.push("/login");
+    }
   } else {
-    localStorage.clear();
-    Route.push("/login");
+    window.location.href = "/login";
   }
 };
 
@@ -182,6 +185,7 @@ export const postWithAuth = (url, data) => {
   };
   // return axios.post(url, data, { headers });
   return new Promise((resolve, reject) => {
+    console.log("🚀 ~ refreshTokenAndSetAuth ~ response:", response);
     axios
       .post(url, data, { headers })
       .then((response) => {
